Changed stlsoft process timer to be toleant instead of asserting
Nate and I are running in to rare but critical errors on OS X only (so
far) where the timers are asserting, claiming their recorded stop times
are happening before their recorded start times. On OS X this breaks
down to either the system call returning bad values or stop getting
called before start. I find the latter to be highly improbable given the
fact that Soar is single threaded and the incidence is unpredictable and
rare (somewhere around 1 in about 500 runs).

I'd love to figure this out and contribute data to stlsoft, but, for
now, I've made the code tolerant instead of asserting (defaulting to
zero in the case that would have assertted) so that research can be
done.