Fixed an epmem bug introduced in r13408.
This bug only occurs if epmem is queried with an LTI, and if the LTI was promoted (stored into smem) in the middle of its most recent interval. Originally, the SQL query uses the promotion time as a constraint; however, since the WME existed before the LTI's promotion, the SQL query fails to find the interval.

The obvious thing to do here is to remove that constraint.

But that causes other problems due to the fact that the epmem search algorithm uses a priority queue to iterate over the intervals endpoints. The priority for a start endpoint of this interval is usually the endpoint itself - except that for an LTI, we don't want to get results before the LTI's promotion; a query for an LTI, constrained to be before its promotion, should result in failure. Note that it's *failure*, not *zero scored episode*, which is what you would get if the LTI appeared and disappeared after the before constraint.

The solution here is to detect when this case happens - if the LTI is promoted in its most recent interval - and set the priority to its promotion time instead of the interval start time.

A couple other small changes are included in this commit, for expanding on the debugging code for WME storage in epmem (unrelated to the above-described bug).